requirements.txt
Pillow==10.3.0
pip==24.0
pywin32==306
typing==3.7.4.3
urllib3==2.2.1
watchdog==4.0.1